requirement.txt
Django>=5.1,<6
djangorestframework>=3.15
psycopg2-binary
Celery
redis
requests
beautifulsoup4
python-whois
dnspython
python-dotenv
drf-yasg            